+++ TDEngine2/include/scene/components/CDirectionalLight.h
@@ -8,11 +8,21 @@
 
 
 #include "ILight.h"
-#include "CBaseLight.h"
+#include "../../ecs/CBaseComponent.h"
 
 
 namespace TDEngine2
 {
+	CLASS_META(SECTION = ecs, flags = SERIALIZE_MARKED_ONLY_FIELDS)
+	struct TDirectionalLightComponentData
+	{
+		FIELD_META(name = color) TColor32F mColor = TColorUtils::mWhite;
+		FIELD_META(name = intensity) F32   mIntensity = 1.0f;
+
+		TDE2_DECLARE_COMPONENT_META(TDirectionalLightComponentData);
+	};
+
+
 	/*!
 		\brief A factory function for creation objects of CDirectionalLight's type
 
@@ -30,7 +40,7 @@ namespace TDEngine2
 		\brief The class represents directed light source which is an analogue of sun light 
 	*/
 
-	class CDirectionalLight : public CBaseLight, public IDirectionalLight, public CPoolMemoryAllocPolicy<CDirectionalLight, 1 << 20>
+	class CDirectionalLight : public CBaseComponentT<CDirectionalLight, TDirectionalLightComponentData>
 	{
 		public:
 			friend TDE2_API IComponent* CreateDirectionalLight(E_RESULT_CODE&);
@@ -38,33 +48,10 @@ namespace TDEngine2
 			TDE2_REGISTER_COMPONENT_TYPE(CDirectionalLight)
 
 			/*!
-				\brief The method deserializes object's state from given reader
-
-				\param[in, out] pReader An input stream of data that contains information about the object
-
-				\return RC_OK if everything went ok, or some other code, which describes an error
+				\return The method returns type name (lowercase is preffered)
 			*/
 
-			TDE2_API E_RESULT_CODE Load(IArchiveReader* pReader) override;
-
-			/*!
-				\brief The method serializes object's state into given stream
-
-				\param[in, out] pWriter An output stream of data that writes information about the object
-
-				\return RC_OK if everything went ok, or some other code, which describes an error
-			*/
-
-			TDE2_API E_RESULT_CODE Save(IArchiveWriter* pWriter) override;
-
-			/*!
-				\brief The method creates a new deep copy of the instance and returns a smart pointer to it.
-				The original state of the object stays the same
-
-				\param[in] pDestObject A valid pointer to an object which the properties will be assigned into
-			*/
-
-			TDE2_API E_RESULT_CODE Clone(IComponent*& pDestObject) const override;
+			TDE2_API const std::string& GetTypeName() const override;
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CDirectionalLight)
 	};