+++ TDEngine2/source/scene/components/CPointLight.cpp
@@ -1,74 +1,23 @@
 #include "../../include/scene/components/CPointLight.h"
+#define META_EXPORT_ECS_SECTION
+#include "../../include/metadata.h"
 
 
 namespace TDEngine2
 {
 	TDE2_REGISTER_COMPONENT_FACTORY(CreatePointLightFactory)
+	TDE2_DEFINE_COMPONENT_META(TPointLightComponentData);
 
 
 	CPointLight::CPointLight() :
-		CBaseLight()
+		CBaseComponentT()
 	{
 	}
 
-	E_RESULT_CODE CPointLight::Load(IArchiveReader* pReader)
+	const std::string& CPointLight::GetTypeName() const
 	{
-		if (!pReader)
-		{
-			return RC_FAIL;
-		}
-
-		mRange = pReader->GetFloat("range");
-
-		return RC_OK;
-	}
-
-	E_RESULT_CODE CPointLight::Save(IArchiveWriter* pWriter)
-	{
-		if (!pWriter)
-		{
-			return RC_FAIL;
-		}
-
-		pWriter->BeginGroup("component");
-		{
-			pWriter->SetUInt32("type_id", static_cast<U32>(CPointLight::GetTypeId()));
-			pWriter->SetFloat("range", mRange);
-		}
-		pWriter->EndGroup();
-
-		return RC_OK;
-	}
-
-	E_RESULT_CODE CPointLight::Clone(IComponent*& pDestObject) const
-	{
-		if (auto pComponent = dynamic_cast<CPointLight*>(pDestObject))
-		{
-			pComponent->mColor = mColor;
-			pComponent->mIntensity = mIntensity;
-			pComponent->mRange = mRange;
-
-			return RC_OK;
-		}
-
-		return RC_FAIL;
-	}
-
-	E_RESULT_CODE CPointLight::SetRange(F32 range)
-	{
-		if (range < 0.0f)
-		{
-			return RC_INVALID_ARGS;
-		}
-
-		mRange = range;
-
-		return RC_OK;
-	}
-
-	F32 CPointLight::GetRange() const
-	{
-		return mRange;
+		static const std::string id{ "point_light" };
+		return id;
 	}
 	
 
@@ -100,9 +49,10 @@ namespace TDEngine2
 			return RC_INVALID_ARGS;
 		}
 
-		pComponent->SetColor(params.mColor);
-		pComponent->SetIntensity(params.mIntensity);
-		pComponent->SetRange(params.mRange);
+		TPointLightComponentData& pointLightData = pComponent->GetData();
+		pointLightData.mColor     = params.mColor;
+		pointLightData.mIntensity = params.mIntensity;
+		pointLightData.mRange     = params.mRange;
 
 		return RC_OK;
 	}