+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.34] 2026-07-18
+## [0.6.34] 2026-07-19
 
 ### Added
 
@@ -29,6 +29,8 @@ automated serialization/deserialization of components.
 - Incorrect work of **CBaseComponentT<T>::GetPropertyFactoriesTable()** was fixed.
 
 - Now only one type of a component for a camera is used which is **CCamera** that can be configured for both types of projection. **COrthoCamera** and **CPerspectiveCamera** components were removed.
+
+- **ILight**, **IDirectionalLight**, **IPointLight**, **ISpotLight** interfaces were removed. Now all light types are inherited from **CBaseComponentT<TComponentType, TComponentDataType>**.
 
 ### Fixed